pyproject.toml
[project]
name = "rag-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.8",
    "google-genai>=1.63.0",
    "inngest>=0.5.15",
    "llama-index-core>=0.14.14",
    "llama-index-embeddings-google-genai>=0.3.2",
    "llama-index-llms-google-genai>=0.8.7",
    "llama-index-readers-file>=0.5.6",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "streamlit>=1.54.0",
    "uvicorn>=0.40.0",
]
